refactor: rename utils to helpers
