+++ frontend/src/helpers/FormatDate/formatDate.ts
@@ -1,6 +1,6 @@
 import { format } from "date-fns";
 
-import { DateFormat } from "@/utils/FormatDate/formatDate.types";
+import { DateFormat } from "@/helpers/FormatDate/formatDate.types";
 
 export const formatDate = (date: number, dateFormat: DateFormat): string => {
   return format(date, dateFormat);
