+++ frontend/src/helpers/formatMessage.ts
